yarn add bootstrap @popperjs/core sass
This commit gets bootstrap to work (css & js).

The @popper/core is a bootstrap dependency for
JavaScript.

sass is also installed because Vite uses it as
a preprocessor for .scss files.

Expose a user_signed_in? helper from
UserAuthentication

include UserAuthentication in the main
ApplicationController because we need helpers
like user_signed_in? on some pages where we
don't enforce user authentication but still
need to show custom-user-elements for logged-in
users.