TLDR; Add the forgot password functionality
Configs touched:

Set default_url_options in development and test to
get the links in the mailers working.

Set the active_job.queue_adapter in test to :test
so .deliver_later in mailers work in tests.

I haven't set a proper queue adapter for prod or
development.

Disable the <div class="field_with_errors"> wrapper
in forms with models.

Forgot password details:

We redirect and send a successful flash message
after a reset password request to avoid enumeration.

Leverage the existing UserToken model to store a
hashed version of a token and use the raw version
in emails. When the link with the token is visited,
we verify the token by running it against the same
hash function. We also compare the user's email
(the one associated with the token via the user_id
reference) to the token's sent_to value to ensure
that the token used isn't stale. The 'sent_to' is
set every time we issue a password reset token.

Move User#create_session_token! to UserToken
The UserToken model should handle it like how it
handles password_reset token creation.

Move User#find_by_session_token to UserToken and rename it
Rename the get_user_by_reset_password_token to find_
user_by_reset_password_token just so it conforms to
ActiveRecord's methods.