Add the user-change-email functionality
When a user wants to change his email, we send a
confirmation link to the new email.

The token has a reference to the user, the user's
current email (in the context) and the new email
(sent_to). When the same user visits the link,
we will change the user's email to the new email
(the one in sent_to). Users can only have access
to the tokens they own since we require them to
be logged in and we scope the query.

All change email tokens are deleted after a change.