Soft deleting a recurring expense, and deleting all instances
Soft deleted ones don't show up in the calendar and don't count in
the breakdown.

There's also an option to delete all instances but it sucks because
we're scoping the query by the `purpose` which isn't reliable at all
but works for now. I'm considering using a different model to store
recurring expense data and all the `Expense` "instances" of that
model have a reference to it. From there we can probably scope by
the FK which is better.